refactor: mutualise l'envoi de messages dans un fichier
